Version resources with pid/time.
Instead of versioning everything with a monotonic counter, use a
resource version based on a 16-bit hash of the current process id (more
on that below) and the current unix timestamp as a 48-bit integer.

In most environments, the current timestamp is monotonic enough to be a
reasonable starting place while still keeping ezbake stateless. Given
the pace of kube updates and the coalescing we do, milliseconds is fine
grained enough. 48 bits of unix time is a little over 8,000 years. If
ezbake is still running then, I'm sorry.

The pid identifier is a hash of the current process id and the
`POD_NAME` environment variable. The combination of these two things
should be unique enough inside and outside kubernetes to avoid having
two processes generate the same identifier.