CLI and Error cleanup (#11)
Cleaned up the CLI args and rennamed them to --namespace and --all-namespaces since that seems simpler. The two args are now mutually exclusive with each other.

There are now reasonable error messages printed out instead of panics when clients disconnect, something bad happens with kube watches, or invalid data gets sent to an ADS connection.

The server also knows how to detect that HTTPRoutes are not installed and logs a warning, but continues on otherwise.