Rewrite the ezbake ingest loop as a single task
The first iteration of ezbake assumed a nice, simple mapping of Kube
resources to xDS. this stopped being true with the interaction between
HTTPRoutes and Services - there are now multiple ways to generate
Listeners, and they interact with each other. Making everything
serialized means we can index and share state when reacting to HTTPRoute
and Service changes, and be a little bit saner. As a bonus, this update
now includes tests.

This update lumps EndpointSlice updates in with the rest of the ingest
tasks, even though it still appears to be independent. If this becomes a
performance bottleneck later, it can move into its own task again.